Fix fprintf warning on Windows under clang.
See issue 21 for more details:
https://code.google.com/p/cld2/issues/detail?id=21



git-svn-id: https://cld2.googlecode.com/svn/trunk@167 b252ecd4-b096-bf77-eb8e-91563289f87e

+++ internal/offsetmap.cc
@@ -79,7 +79,7 @@ void OffsetMap::Printmap(const char* filename) {
   }
 
   Flush();    // Make sure any pending entry gets printed
-  fprintf(fout, "Offsetmap: %ld bytes\n", diffs_.size());
+  fprintf(fout, "Offsetmap: %d bytes\n", static_cast<int>(diffs_.size()));
   for (int i = 0; i < static_cast<int>(diffs_.size()); ++i) {
     fprintf(fout, "%c%02d ", "&=+-"[OpPart(diffs_[i])], LenPart(diffs_[i]));
     if ((i % 20) == 19) {fprintf(fout, "\n");}
